requirements.txt
# YOLO 训练与部署依赖包

# 基础依赖
torch>=2.0.0
torchvision>=0.15.0

# YOLO相关
ultralytics>=8.0.0  # YOLOv8

# 计算机视觉
opencv-python>=4.8.0
pillow>=10.0.0
matplotlib>=3.7.0
numpy>=1.24.0
scipy>=1.10.0

# 数据处理
pandas>=2.0.0
pyyaml>=6.0

# 部署相关
onnx>=1.14.0
onnxruntime>=1.15.0  # CPU推理
onnxruntime-gpu>=1.15.0  # GPU推理

# 可选：TensorRT支持
# tensorrt>=8.6.0
# pycuda>=2022.2

# 可选：边缘设备部署
# tflite-runtime>=2.14.0

# 开发工具
ipython>=8.14.0
jupyter>=1.0.0
notebook>=7.0.0

# 测试
pytest>=7.4.0